Code refactoring and general reworking

Approved-by: fdesimone

+++ paw-webapp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/EmailService.java
@@ -8,9 +8,9 @@ import java.util.Locale;
 
 public interface EmailService {
 
-    void purchase(String buyerEmail, String buyer, Product product, int amount, float price,
+    void purchase(String buyerEmail, String buyer, Product product, int amount, Integer price,
                   String sellerName, String sellerPhone, String sellerMail, Locale locale);
-    void itemsold(String sellerEmail, String seller, Product product, int amount, float price,
+    void itemsold(String sellerEmail, String seller, Product product, int amount, Integer price,
                   String buyerName, String buyerEmail, String buyerMessage, Locale locale);
 
     void registration(User user, Locale locale);
